Model.RESTDatatypes

Description

Module contains all datatypes used by the REST service and functions to parse all ready members and strings

Documentation

data Login

All REST Datatypes which can be packed into json und unpacked

Constructors

Login

Fields

loginMail :: !Text

loginPassword :: !Text

type LoginId = Key Login

data Session

Constructors

Session

Fields

sessionKey :: !Int

type SessionId = Key Session

data Member

Constructors

Member

Fields

memberName :: !Text

memberSurName :: !Text

memberBirthDay :: !Int

memberBirthMonth :: !Int

memberBirthYear :: !Int

memberExamationDay :: !Int

memberExamationMonth :: !Int

memberExamationYear :: !Int

memberInstructionCheck :: !Int

memberExerciseCheck :: !Int

type MemberId = Key Member

data Memberlist

Constructors

Memberlist

Fields

memberlistMembers :: ![Member]

type MemberlistId = Key Memberlist

data Membersupdate

Constructors

Membersupdate

Fields

membersupdateType :: !Text

membersupdateMemberSurNames :: ![Text]

membersupdateMemberNames :: ![Text]

type MembersupdateId = Key Membersupdate

data Appointment

Constructors

Appointment

Fields

appointmentTitle :: !Text

appointmentType :: !Text

appointmentDay :: !Int

appointmentMonth :: !Int

appointmentYear :: !Int

appointmentHour :: !Int

appointmentMinute :: !Int

appointmentMembers :: ![Text]

type AppointmentId = Key Appointment

data Appointmentlist

Constructors

Appointmentlist

Fields

appointmentlistAppointments :: ![Appointment]

type AppointmentlistId = Key Appointmentlist

migrateAll :: Migration

localTime

Arguments

:: IO (Integer, Int, Int)
Do IO Return current Date

Get Current Time in (Year, Month, Day)

membersReady

Arguments

:: [Entity Member]
List of Members

-> (Integer, Int, Int)
Current Date

-> [Entity Member]
Return Members which joined all importend appointments

Get Return Members which joined all importend appointments

membersReadyList :: [Entity Member] -> (Integer, Int, Int) -> [Bool]

Get members which joined all importend appointments

appointmentsInFuture

Arguments

:: [Entity Appointment]
List of Appointments

-> (Integer, Int, Int)
Current Date

-> [Entity Appointment]
Appointments in future

Get appointments in Future

appointmentInFuture

Arguments

:: Entity Appointment
Appointment to check

-> (Int, Int, Integer)
Current Date

-> Bool
Flag if date is in future

Check if appointment is in future

membersReadyFilter

Arguments

:: Entity Member
Member

-> (Int, Int, Integer)
Current Date

-> Bool
Flag if member is ready

Check if member joined all importend appointments

dateToString

Arguments

:: (Int, Int, Int)
Date to convert

-> String
return converted date

Convert Date to String

timeToString

Arguments

:: (Int, Int)
Time to convert

-> String
Converted time

Convert Time to String